Line: allow lines to be drawn in negative ranges

+++ GfxPrimitives/Line.cpp
@@ -8,11 +8,11 @@ Line::Line():
     Line(0, 0, 0, 0, CRGB(0, 0, 0))
 {}
 
-Line::Line(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1):
+Line::Line(int16_t x0, int16_t y0, int16_t x1, int16_t y1):
     Line(x0, y0, x1, y1, CRGB(0, 0, 0))
 {}
 
-Line::Line(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, const CRGB &color):
+Line::Line(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const CRGB &color):
     m_x0(x0), m_y0(y0), m_x1(x1), m_y1(y1), m_color(color), m_gradient(CRGB(0, 0, 0))
 {
     update();
@@ -22,17 +22,17 @@ Line::Line(Point p0, Point p1):
     Line(p0.x, p0.y, p1.x, p1.y, CRGB(0, 0, 0))
 {}
 
-uint8_t Line::getX0() const
+int16_t Line::getX0() const
 {
     return m_x0;
 }
 
-uint8_t Line::getY0() const
+int16_t Line::getY0() const
 {
     return m_y0;
 }
 
-Line& Line::setStart(uint8_t x, uint8_t y)
+Line& Line::setStart(int16_t x, int16_t y)
 {
     m_x0 = x;
     m_y0 = y;
@@ -41,17 +41,17 @@ Line& Line::setStart(uint8_t x, uint8_t y)
     return *this;
 }
 
-uint8_t Line::getX1() const
+int16_t Line::getX1() const
 {
     return m_x1;
 }
 
-uint8_t Line::getY1() const
+int16_t Line::getY1() const
 {
     return m_y1;
 }
 
-Line& Line::setEnd(uint8_t x, uint8_t y)
+Line& Line::setEnd(int16_t x, int16_t y)
 {
     m_x1 = x;
     m_y1 = y;
@@ -90,10 +90,10 @@ Line & Line::setGradient(const CRGBPalette16 &gradient)
 
 void Line::update()
 {
-    uint8_t x0 = m_x0;
-    uint8_t y0 = m_y0;
-    uint8_t x1= m_x1;
-    uint8_t y1 = m_y1;
+    int16_t x0 = m_x0;
+    int16_t y0 = m_y0;
+    int16_t x1= m_x1;
+    int16_t y1 = m_y1;
 
     int16_t steep = abs(y1 - y0) > abs(x1 - x0);
     if (steep) {
@@ -119,7 +119,7 @@ void Line::update()
         ystep = -1;
     }
 
-    auto colorAtX = [this, x0, x1](uint8_t x) {
+    auto colorAtX = [this, x0, x1](int16_t x) {
         if (!m_useGradient) {
             return m_color;
         }
