+++ GfxPrimitives/Line.h
@@ -7,17 +7,17 @@ class Line : public GfxPrimitive
 {
 public:
     Line();
-    Line(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1);
-    Line(uint8_t x0, uint8_t y0, uint8_t x1, uint8_t y1, const CRGB &color);
+    Line(int16_t x0, int16_t y0, int16_t x1, int16_t y1);
+    Line(int16_t x0, int16_t y0, int16_t x1, int16_t y1, const CRGB &color);
     Line(Point p0, Point p1);
 
-    uint8_t getX0() const;
-    uint8_t getY0() const;
-    Line& setStart(uint8_t x, uint8_t y);
+    int16_t getX0() const;
+    int16_t getY0() const;
+    Line& setStart(int16_t x, int16_t y);
 
-    uint8_t getX1() const;
-    uint8_t getY1() const;
-    Line& setEnd(uint8_t x, uint8_t y);
+    int16_t getX1() const;
+    int16_t getY1() const;
+    Line& setEnd(int16_t x, int16_t y);
 
     const CRGB &getColor() const;
     Line& setColor(const CRGB &color);
@@ -26,10 +26,10 @@ public:
     Line& setGradient(const CRGBPalette16 &gradient);
 
 protected:
-    uint8_t m_x0;
-    uint8_t m_y0;
-    uint8_t m_x1;
-    uint8_t m_y1;
+    int16_t m_x0;
+    int16_t m_y0;
+    int16_t m_x1;
+    int16_t m_y1;
 
     CRGB m_color;
